% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_onset.R
\name{defineOnset}
\alias{defineOnset}
\title{Define critical onset in iChart}
\usage{
defineOnset(iChart, critonset = 300, includeAways = FALSE)
}
\arguments{
\item{iChart}{the iChart}

\item{critonset}{a number}

\item{includeAways}{logical}
}
\description{
This function allows you to set the critical onset for the analysis window in the iChart
}
\examples{
defineOnset(iChart = "iChart", critonset = 300, includeAways=F)
}

