% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ichart.R
\name{filteriChart}
\alias{filteriChart}
\title{Filter iChart}
\usage{
filteriChart(iChart, minRT, maxRT, maxfirstgap, maxlonggap)
}
\arguments{
\item{iChart}{the iChart}

\item{minRT}{lower boundary of anlaysis window}

\item{maxRT}{upper boundary of anlaysis window}

\item{maxfirstgap}{max first gap length}

\item{maxlonggap}{max gap length}
}
\description{
This function allows you to filter out bad trials from your iChart
}

