% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_frame_convert.R
\name{framesToMs}
\alias{framesToMs}
\title{Convert frames to ms in an iChart}
\usage{
framesToMs(value)
}
\arguments{
\item{numeric}{}
}
\description{
This function allows you to convert frames to ms for raw eye movement data from the looking-while-listening task
}
\examples{
framesToMs(value)
}

