% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwl.R
\docType{package}
\name{LWL-package}
\alias{LWL}
\alias{LWL-package}
\title{LWL: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
