% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_frame_convert.R
\name{msToFrames}
\alias{msToFrames}
\title{Convert ms to frames in an iChart}
\usage{
msToFrames(value)
}
\arguments{
\item{numeric}{}
}
\description{
This function allows you to convert ms to frames for raw eye movement data from the looking-while-listening task
}
\examples{
msToFrames(value)
}

