% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readiChart.R
\name{readiChart}
\alias{readiChart}
\title{read iChart}
\usage{
readiChart(iChartFile)
}
\arguments{
\item{iChartFile}{}
}
\description{
This function allows you to read in raw eye movement data from the looking-while-listening task
}
\examples{
readiChart(iChartFile = "iChart")
}
\keyword{iChart}

